+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [v0.28.0] - 2021-10-18
 ### Added
 - mongo: adds support for `mongodb+srv` connection strings.
 - mongo: binds in a default TLS Config if `ssl=true` and a TLS config has not been provided.
@@ -12,7 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - mongo: implements `storage.Expirer` interface to enable TTL based expiry on tokens.
 
 ### Changed
-- mongo: migrated use of `isDup(err)` to `mongo.IsDuplicateKeyError(err)`.
+- mongo: migrated internal use of `isDup(err)` to `mongo.IsDuplicateKeyError(err)`.
 
 ### Removed
 - mongo: removed internal `isDup(err)` function.
@@ -653,6 +654,7 @@ clear out the password field before sending the response.
 - General pre-release!
 
 [Unreleased]: https://github.com/matthewhartstonge/storage/tree/master
+[v0.28.0]: https://github.com/matthewhartstonge/storage/tree/v0.28.0
 [v0.27.0]: https://github.com/matthewhartstonge/storage/tree/v0.27.0
 [v0.26.0]: https://github.com/matthewhartstonge/storage/tree/v0.26.0
 [v0.25.1]: https://github.com/matthewhartstonge/storage/tree/v0.25.1